Potential level cap is now restricted to 2 byte size (65535) instead of 1 byte (255), currently capped at Lvl 1000

+++ include/battle.h
@@ -897,7 +897,7 @@ struct BattleScripting
     u8 savedBattler;
     u8 reshowMainState;
     u8 reshowHelperState;
-    u8 levelUpHP;
+    u16 levelUpHP;
     u8 windowsType; // B_WIN_TYPE_*
     u8 multiplayerId;
     u8 specialTrainerBattleType;
