+++ src/trade.c
@@ -2013,7 +2013,7 @@ static void PrintPartyNicknames(u8 whichParty)
 
 static void PrintLevelAndGender(u8 whichParty, u8 monIdx, u8 x, u8 y, u8 width, u8 height)
 {
-    u8 level;
+    u16 level;
     u32 symbolTile;
     u8 gender;
     u8 nickname[POKEMON_NAME_LENGTH + 1];
@@ -4532,7 +4532,7 @@ static void BufferInGameTradeMonName(void)
 static void CreateInGameTradePokemonInternal(u8 whichPlayerMon, u8 whichInGameTrade)
 {
     const struct InGameTrade *inGameTrade = &sIngameTrades[whichInGameTrade];
-    u8 level = GetMonData(&gPlayerParty[whichPlayerMon], MON_DATA_LEVEL);
+    u16 level = GetMonData(&gPlayerParty[whichPlayerMon], MON_DATA_LEVEL);
 
     struct Mail mail;
     u8 metLocation = METLOC_IN_GAME_TRADE;
