+++ src/tv.c
@@ -67,7 +67,7 @@ COMMON_DATA s8 sCurTVShowSlot = 0;
 COMMON_DATA u16 sTV_SecretBaseVisitMovesTemp[8] = {0};
 COMMON_DATA u8 sTV_DecorationsBuffer[DECOR_MAX_SECRET_BASE] = {0};
 COMMON_DATA struct {
-    u8 level;
+    u16 level;
     u16 species;
     u16 move;
 } sTV_SecretBaseVisitMonsTemp[10] = {0};
@@ -191,7 +191,7 @@ static void DoTVShowLilycoveContestLady(void);
 static const struct {
     u16 species;
     u16 moves[MAX_MON_MOVES];
-    u8 level;
+    u16 level;
     u8 location;
 } sPokeOutbreakSpeciesList[] = {
     {
